Get J-Stage Articles List

Retrieves a list of articles published on J-STAGE, including detailed information such as publication date and bibliographic metadata.

Usage

get_jstage_articles(
  pubyearfrom = NA,
  pubyearto = NA,
  material = "",
  article = "",
  author = "",
  affil = "",
  keyword = "",
  abst = "",
  text = "",
  issn = "",
  cdjournal = "",
  sortflg = NA,
  vol = NA,
  no = NA,
  start = NA,
  count = NA,
  sep = "\n",
  lang = "ja"
)

Arguments

pubyearfrom

An integer specifying the starting publication year (in YYYY format).

pubyearto

An integer specifying the ending publication year (in YYYY format).

material

A character string specifying the material name (partial match search, case insensitive).

article

A character string specifying the article title (partial match search, case insensitive).

author

A character string specifying the author name (partial match search, case insensitive).

affil

A character string specifying the affiliation (partial match search, case insensitive).

keyword

A character string specifying the keyword (partial match search, case insensitive).

abst

A character string specifying the abstract (partial match search, case insensitive).

text

A character string specifying the full text (partial match search, case insensitive).

issn

A character string specifying the ISSN (exact match search in XXXX-XXXX format).

cdjournal

A character string specifying the journal code.

sortflg

An integer specifying the sort flag: 1 for score order, 2 for volume, issue, and page order (default is 1).

vol

An integer specifying the volume (exact match).

no

An integer specifying the issue number (exact match).

start

An integer specifying the starting index for the search results.

count

An integer specifying the number of search results to retrieve (up to 1,000).

sep

A character string to separate multiple authors' names. Default is "\n".

lang

A character string specifying the language for column names: "ja" for Japanese (default is "ja").

Value

A list containing metadata and entry data frames with the search results.