Get J-Stage Volumes and Issues List

Retrieves a list of volumes and issues published on J-STAGE.

Usage

get_jstage_volumes(
  pubyearfrom = NA,
  pubyearto = NA,
  material = "",
  issn = "",
  cdjournal = "",
  volorder = NA,
  lang = "ja"
)

Arguments

pubyearfrom

An integer specifying the starting publication year (in YYYY format).

pubyearto

An integer specifying the ending publication year (in YYYY format).

material

A character string specifying the material name (exact match search).

issn

A character string specifying the ISSN (exact match search in XXXX-XXXX format).

cdjournal

A character string specifying the journal code.

volorder

An integer specifying the order of volumes: 1 for ascending, 2 for descending (default is 1).

lang

A character string specifying the language for column names: "ja" for Japanese (default is "ja").

Value

A list containing metadata and entry data frames with the search results.