Scrape J-STAGE References

Scrapes all available references from a J-STAGE article web page.

Usage

jstage_references(url, depth = 1, citedby = FALSE, wait = 1, quiet = TRUE)

Arguments

url

The URL or DOI of the J-STAGE article web page.

depth

Integer. The depth to which references should be scraped. For example, a depth of 2 means the references of the specified paper and recursively the references of those references are scraped, and so on.

citedby

Logical. Set to TRUE to retrieve the list of cited-by references instead of the reference list. When set to TRUE, depth is automatically set to 1.

wait

Numeric. The number of seconds to wait between each request to reduce server load.

quiet

Logical. Set to FALSE to display information about the progress.

Value

A data frame with the DOI of each reference.